refactor: 리포지토리를 도메인별로 정리
