+++ src/main/java/team1/housework/group/repository/housework/HouseWorkRepository.java
@@ -1,4 +1,4 @@
-package team1.housework.group.repository;
+package team1.housework.group.repository.housework;
 
 import java.time.LocalDate;
 import java.util.List;
