+++ src/main/java/team1/housework/group/repository/houseworkmember/HouseWorkMemberRepository.java
@@ -1,4 +1,4 @@
-package team1.housework.group.repository;
+package team1.housework.group.repository.houseworkmember;
 
 import org.springframework.data.jpa.repository.JpaRepository;
