+++ src/main/java/team1/housework/group/repository/houseworktag/HouseWorkTagCustomRepository.java
@@ -1,4 +1,4 @@
-package team1.housework.group.repository;
+package team1.housework.group.repository.houseworktag;
 
 import java.util.List;
