+++ src/main/java/team1/housework/group/repository/tag/TagCustomRepositoryImpl.java
@@ -1,4 +1,4 @@
-package team1.housework.group.repository;
+package team1.housework.group.repository.tag;
 
 import static team1.housework.group.entity.QTag.*;
