+++ src/main/java/team1/housework/group/service/GroupService.java
@@ -22,11 +22,11 @@ import team1.housework.group.entity.Place;
 import team1.housework.group.entity.Tag;
 import team1.housework.group.repository.GroupMemberRepository;
 import team1.housework.group.repository.GroupRepository;
-import team1.housework.group.repository.HouseWorkMemberRepository;
-import team1.housework.group.repository.HouseWorkRepository;
-import team1.housework.group.repository.HouseWorkTagRepository;
 import team1.housework.group.repository.PlaceRepository;
-import team1.housework.group.repository.TagRepository;
+import team1.housework.group.repository.housework.HouseWorkRepository;
+import team1.housework.group.repository.houseworkmember.HouseWorkMemberRepository;
+import team1.housework.group.repository.houseworktag.HouseWorkTagRepository;
+import team1.housework.group.repository.tag.TagRepository;
 import team1.housework.group.service.dto.EnterRequest;
 import team1.housework.group.service.dto.EnterResponse;
 import team1.housework.group.service.dto.GroupRequest;
